requirements.txt
psycopg2
falcon==0.1.8
gunicorn
requests
pyyaml
